refactor(toast): move toast files to providers/ and hooks/
