--- src/contexts/index.ts
@@ -1,2 +1,0 @@
-export { ToastProvider } from "./ToastProvider";
-export { useToast } from "./useToast";
