+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 export * from "./useStateContext";
 export * from "./useNavigateWithParams";
+export * from "./useToast";
 export * from "./useWallet";
 export * from "./useTransactionExecutor";
 export * from "./useCanonGuardConfig";
